Move LB code to /LoadBalancing/ from /EngineClasses/
